+++ src/main/java/pigcart/puddleflood/mixin/FlowingFluidMixin.java
@@ -4,21 +4,23 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.material.FlowingFluid;
 import net.minecraft.world.level.material.FluidState;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import pigcart.puddleflood.BlockPlacementUtil;
-import pigcart.puddleflood.PuddleFlood;
-import pigcart.puddleflood.config.ConfigManager;
+import net.minecraft.world.level.Level;
 //?>=1.21.9{
 /*import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.block.state.BlockState;
-*///?} else {
-import net.minecraft.world.level.Level;
-//?}
+*///?}
 
 @Mixin(FlowingFluid.class)
 public abstract class FlowingFluidMixin {
+
+    @Unique
+    public void placePuddle(Level level, BlockPos pos) {
+        // overridden in WaterFluidMixin
+    }
 
     @Inject(method = "spread", at = @At("HEAD"))
     protected void spread(
@@ -28,8 +30,6 @@ public abstract class FlowingFluidMixin {
             FluidState fluidState,
             CallbackInfo ci
     ) {
-        if (ConfigManager.config.doFlowingWaterPuddles && level.random.nextBoolean() && BlockPlacementUtil.isUnobstructedFlatSurface(level, pos)) {
-            level.setBlockAndUpdate(pos, PuddleFlood.PUDDLE_BLOCK.getStateAt(pos, level));
-        }
+        placePuddle(level, pos);
     }
 }